refactor: structure and types (#59)
* move util file

* accept zod schema in `useMatchesData` and remove `useActiveNavigation`

* move logic out of data folder to utils

* change property name

* refactor: structure, types
