+++ app/hooks/useUser.ts
@@ -1,0 +1,11 @@
+import useOptionalUser from './useOptionalUser'
+
+export default function useUser() {
+  const maybeUser = useOptionalUser()
+  if (!maybeUser) {
+    throw new Error(
+      'No user found in root loader, but user is required by useUser. If user is optional, try useOptionalUser instead.'
+    )
+  }
+  return maybeUser
+}
