+++ app/models/character.server.ts
@@ -1,7 +1,5 @@
-// import invariant from 'tiny-invariant'
 import prisma from '~/db.server'
-// import type * as Utils from '~/utils'
-// import * as UserModel from './user.server'
+import type * as CharacterType from '~/types/character'
 
 export async function getUserCharacters({ accountId }: { accountId: string }) {
   return prisma.userCharacter.findMany({
@@ -9,7 +7,7 @@ export async function getUserCharacters({ accountId }: { accountId: string }) {
       ownerId: accountId,
     },
     select: {
-      characterName: true,
+      name: true,
       level: true,
       ascension: true,
       normalAttack: true,
@@ -23,13 +21,13 @@ export async function getUserCharacter({
   name,
   accountId,
 }: {
-  name: string
+  name: CharacterType.Name
   accountId: string
 }) {
   return prisma.userCharacter.findFirst({
     where: {
       ownerId: accountId,
-      characterName: name,
+      name,
     },
     select: {
       level: true,
@@ -46,7 +44,7 @@ export async function upsertCharacter({
   progression,
   accountId,
 }: {
-  name: string
+  name: CharacterType.Name
   progression: {
     level: number
     ascension?: number
@@ -68,7 +66,7 @@ export async function upsertCharacter({
       const characters = await tx.userCharacter.findMany({
         where: {
           ownerId: accountId,
-          characterName: {
+          name: {
             contains: 'Traveler',
           },
         },
@@ -83,19 +81,19 @@ export async function upsertCharacter({
           data: [
             {
               ownerId: accountId,
-              characterName: name,
+              name,
               ...defaultProgression,
               ...progression,
             },
             {
               ownerId: accountId,
-              characterName: names[0],
+              name: names[0],
               ...progression,
               ...defaultProgression,
             },
             {
               ownerId: accountId,
-              characterName: names[1],
+              name: names[1],
               ...progression,
               ...defaultProgression,
             },
@@ -103,14 +101,14 @@ export async function upsertCharacter({
         })
       } else {
         await tx.userCharacter.update({
-          where: { id: characters.find((c) => c.characterName === name)!.id },
+          where: { id: characters.find((c) => c.name === name)!.id },
           data: { ...progression },
         })
 
         await tx.userCharacter.updateMany({
           where: {
             ownerId: accountId,
-            characterName: {
+            name: {
               contains: 'Traveler',
             },
           },
@@ -127,7 +125,7 @@ export async function upsertCharacter({
     const character = await tx.userCharacter.findFirst({
       where: {
         ownerId: accountId,
-        characterName: name,
+        name,
       },
     })
 
@@ -140,7 +138,7 @@ export async function upsertCharacter({
       await tx.userCharacter.create({
         data: {
           ownerId: accountId,
-          characterName: name,
+          name,
           ...defaultProgression,
           ...progression,
         },
@@ -151,7 +149,7 @@ export async function upsertCharacter({
 
 export async function updateUserCharacters(
   data: {
-    characterName: string
+    name: CharacterType.Name
     level: number
     ascension: number
     normalAttack: number
@@ -178,7 +176,7 @@ export async function updateCharacterByInventory({
   materials,
   accountId,
 }: {
-  name: string
+  name: CharacterType.Name
   level?: number
   accountId: string
   materials: {
@@ -205,7 +203,7 @@ export async function updateCharacterByInventory({
       await tx.userCharacter.updateMany({
         where: {
           ownerId: accountId,
-          characterName: {
+          name: {
             contains: name.includes('Traveler') ? 'Traveler' : name,
           },
         },
@@ -229,7 +227,7 @@ export async function updateCharacterByInventory({
       await tx.userCharacter.updateMany({
         where: {
           ownerId: accountId,
-          characterName: name,
+          name,
         },
         data: {
           [toUpdate[kind]]: { increment: 1 },
@@ -240,7 +238,7 @@ export async function updateCharacterByInventory({
     await tx.inventory.updateMany({
       where: {
         ownerId: accountId,
-        itemName: materials[0].name,
+        name: materials[0].name,
       },
       data: {
         quantity: {
@@ -252,7 +250,7 @@ export async function updateCharacterByInventory({
     await tx.inventory.updateMany({
       where: {
         ownerId: accountId,
-        itemName: materials[1].name,
+        name: materials[1].name,
       },
       data: {
         quantity: {
@@ -265,7 +263,7 @@ export async function updateCharacterByInventory({
       await tx.inventory.updateMany({
         where: {
           ownerId: accountId,
-          itemName: materials[2].name,
+          name: materials[2].name,
         },
         data: {
           quantity: {
@@ -279,7 +277,7 @@ export async function updateCharacterByInventory({
       await tx.inventory.updateMany({
         where: {
           ownerId: accountId,
-          itemName: materials[3].name,
+          name: materials[3].name,
         },
         data: {
           quantity: {
