+++ app/routes/_app.alchemy.crafting.$type/_index.tsx
@@ -1,43 +1,48 @@
 import * as RemixNode from '@remix-run/node'
 import * as RemixReact from '@remix-run/react'
 import * as Zod from 'zod'
-import * as ItemData from '~/data/items'
 import * as DB from '~/db.server'
 import * as InventoryModel from '~/models/inventory.server'
 import * as Session from '~/session.server'
 import * as Utils from '~/utils'
+import * as UtilsServer from '~/utils/index.server'
 import ItemList from './ItemList'
+
+export const meta: RemixNode.MetaFunction = () => ({
+  title: `Alchemy Crafting - Traveler Main`,
+  description: `Crafting page is to help you craft items like in game`,
+})
 
 export async function loader({ params, request }: RemixNode.LoaderArgs) {
   const accountId = await Session.requireAccountId(request)
   const type = Zod.enum(['all', 'enhancement', 'ascension', 'talent']).parse(
     params.type
   )
+  const parsedType =
+    type === 'enhancement'
+      ? DB.ItemType.COMMON
+      : type === 'ascension'
+      ? DB.ItemType.ASCENSION_GEM
+      : DB.ItemType.TALENT_BOOK
+
   let itemNames: string[] = []
 
   if (type === 'all') {
-    itemNames = ItemData.getCraftItemNames()
+    itemNames = UtilsServer.Item.getCraftItemNames()
   } else {
-    const parsedType =
-      type === 'enhancement'
-        ? DB.ItemType.COMMON
-        : type === 'ascension'
-        ? DB.ItemType.ASCENSION_GEM
-        : DB.ItemType.TALENT_BOOK
-
-    itemNames = ItemData.getCraftItemNamesByType({
+    itemNames = UtilsServer.Item.getCraftItemNamesByType({
       type: parsedType,
     })
   }
 
   const inventory = await InventoryModel.getRequiredItems({
-    itemNames,
+    names: itemNames,
     accountId,
   })
 
-  const craftable = ItemData.getCraftableItems({
+  const craftable = UtilsServer.Item.getCraftableItems({
     userItems: inventory,
-    type: type === 'all' ? undefined : type,
+    type: type === 'all' ? undefined : parsedType,
   })
 
   return RemixNode.json(craftable)
@@ -48,6 +53,8 @@ export default function AlchemyConvertingPage() {
   const craft = Zod.enum(['all', 'enhancement', 'ascension', 'talent']).parse(
     RemixReact.useParams().type
   )
+
+  console.log(items)
 
   if (
     Utils.hasOwnProperty(items, 'craftable') &&
