+++ app/routes/_app.character.$name.inventory-levelup/_index.tsx
@@ -6,12 +6,11 @@ import * as Zod from 'zod'
 import Button from '~/components/Button'
 import * as Icon from '~/components/Icon'
 import Notification from '~/components/Notification'
-import * as CharacterData from '~/data/characters'
-import * as ItemData from '~/data/items'
 import * as DB from '~/db.server'
 import * as CharacterModel from '~/models/character.server'
 import * as InventoryModel from '~/models/inventory.server'
 import * as Session from '~/session.server'
+import * as UtilsServer from '~/utils/index.server'
 import InventoryLevelUpForm from './InventoryLevelUpForm'
 import ItemWithImage from './ItemWithImage'
 
@@ -38,8 +37,15 @@ interface ActionData {
 
 export async function action({ params, request }: RemixNode.ActionArgs) {
   const accountId = await Session.requireAccountId(request)
+
   const { name } = params
-  invariant(name)
+  const validCharacter = UtilsServer.Character.validateCharacter(name)
+  if (!validCharacter) {
+    throw RemixNode.json(`Character ${name} not found`, {
+      status: 404,
+      statusText: 'Character Not Found',
+    })
+  }
 
   const result = await RemixParamsHelper.getFormData(request, ParamsSchema)
   if (!result.success) {
@@ -84,30 +90,36 @@ export async function action({ params, request }: RemixNode.ActionArgs) {
 
 export async function loader({ params, request }: RemixNode.LoaderArgs) {
   const accountId = await Session.requireAccountId(request)
-  const { name } = params
-  invariant(name)
 
-  const fullItemNames = CharacterData.getItemsToRetrieve(name)
+  const { name } = params
+  if (!UtilsServer.Character.validateCharacter(name)) {
+    throw RemixNode.json(
+      { message: `There is no character with name ${name}` },
+      { status: 404, statusText: 'Character not found' }
+    )
+  }
+
+  const fullItemNames = UtilsServer.Character.getItemsToRetrieve(name)
 
   const [userCharacter, userItems] = await Promise.all([
     CharacterModel.getUserCharacter({ name, accountId }),
-    InventoryModel.getRequiredItems({ itemNames: fullItemNames, accountId }),
+    InventoryModel.getRequiredItems({ names: fullItemNames, accountId }),
   ])
 
-  const fullItems = ItemData.getItemsByNames({
+  const fullItems = UtilsServer.Item.getItemsByNames({
     names: fullItemNames,
     userItems,
   })
-  const unlock = CharacterData.getUnlock(userCharacter?.ascension ?? 0)
+  const unlock = UtilsServer.Character.getUnlock(userCharacter?.ascension ?? 0)
 
-  const data = CharacterData.getCharacterInventoryLevelUpData({
+  const data = UtilsServer.Character.getCharacterInventoryLevelUpData({
     name,
     progression: userCharacter,
     itemNames: fullItemNames,
     userItems: fullItems,
   })
 
-  const ascend = CharacterData.checkAscend({
+  const ascend = UtilsServer.Character.getAscendStatus({
     ascension: userCharacter?.ascension,
     level: userCharacter?.level,
   })
