+++ app/routes/_app.character._index/CharacterGrid.tsx
@@ -1,8 +1,10 @@
-import type * as CharacterData from '~/data/characters'
+import type * as CharacterType from '~/types/character'
 import CharacterGridItem from './CharacterGridItem'
 
 interface Props {
-  characters: CharacterData.Character[]
+  characters: (CharacterType.Character & {
+    progression: CharacterType.Progression
+  })[]
 }
 
 export default function CharacterGrid({ characters }: Props) {
@@ -15,7 +17,7 @@ export default function CharacterGrid({ characters }: Props) {
               name={character.name}
               vision={character.vision}
               rarity={character.rarity}
-              level={character.progression?.level}
+              level={character.progression.level}
             />
           </li>
         ))}
