+++ app/utils/index.server.ts
@@ -1,0 +1,4 @@
+import * as Character from './character.server'
+import * as Item from './item.server'
+
+export { Character, Item }
