+++ app/utils/index.ts
@@ -1,6 +1,3 @@
-import * as RemixReact from '@remix-run/react'
-import * as React from 'react'
-
 const DEFAULT_REDIRECT = '/'
 
 /**
@@ -23,74 +20,6 @@ export function safeRedirect(
   }
 
   return to
-}
-
-/**
- * This base hook is used in other hooks to quickly search for specific data
- * across all loader data using useMatches.
- * @param {string} id The route id
- * @returns {JSON|undefined} The router data or undefined if not found
- */
-export function useMatchesData(
-  id: string
-): Record<string, unknown> | undefined {
-  const matchingRoutes = RemixReact.useMatches()
-  const route = React.useMemo(
-    () => matchingRoutes.find((route) => route.id === id),
-    [matchingRoutes, id]
-  )
-  return route?.data
-}
-
-function isUser(user: any): user is {
-  id: string
-  email: string
-  accounts: {
-    id: string
-    name: string
-  }[]
-} {
-  return user && typeof user === 'object' && typeof user.email === 'string'
-}
-
-export function useOptionalUser():
-  | {
-      id: string
-      email: string
-      accounts: {
-        id: string
-        name: string
-      }[]
-    }
-  | undefined {
-  const data = useMatchesData('root')
-  if (!data || !isUser(data.user)) {
-    return undefined
-  }
-  return data.user
-}
-
-export function useUser(): {
-  id: string
-  email: string
-  accounts: {
-    id: string
-    name: string
-  }[]
-} {
-  const maybeUser = useOptionalUser()
-  if (!maybeUser) {
-    throw new Error(
-      'No user found in root loader, but user is required by useUser. If user is optional, try useOptionalUser instead.'
-    )
-  }
-  return maybeUser
-}
-
-export function useActiveNavigation(to: string): boolean {
-  const { pathname } = RemixReact.useLocation()
-  const resolvedPath = RemixReact.useResolvedPath(to).pathname
-  return resolvedPath === pathname || `${resolvedPath}/` === pathname
 }
 
 // https://fettblog.eu/typescript-hasownproperty/
